+++ modules/ParameterExplorer/iAParamSpatialView.h
@@ -39,14 +39,14 @@ class iAParamSpatialView: public QWidget
 	Q_OBJECT
 public:
 	iAParamSpatialView(iAParamTableView* table, QString const & basePath);
-	void SetImage(int id);
+	void SetImage(size_t id);
 private slots:
 	void SlicerModeButtonClicked(bool checked);
 	void SliceChanged(int slice);
 private:
 	iAParamTableView* m_table;
 	QString m_basePath;
-	QMap<int, vtkSmartPointer<vtkImageData>> m_imageCache;
+	QMap<size_t, vtkSmartPointer<vtkImageData>> m_imageCache;
 	QVector<iAITKIO::ImagePointer> m_loadedImgs; // to stop itk from unloading
 	int m_curMode;
 	int m_sliceNr[3];
